Move configure-store into util, replace with barrel file
